Allow specifying both compiler and makeprg
Previously, Makery ignores the makeprg and errorformat options if a
compiler option is given.

However, there is a use case to setting both compiler and makeprg: you
can make use of the proper errorformat from setting the compiler, but
you want to invoke the makeprg in a different manner (e.g. with
additional args, or through your task runner such as in an npm script).

Makery now allows setting both. errorformat will still be ignored if
compiler is set, since I don't see a use case for doing that yet.