Add initial Projectionist support
Discussion in #5.

As of now, a simple projectionist#query() is done for a "makery" key.
This is expected to contain an object of the same shape as described in
:help makery-path-config.

This is especially convenient because it only entails calling
makery#Setup directly. However, later additions might want to take
advantage of smarter queries, e.g. use "make" and "dispatch" to some
extent?