Change recurse logic to match vim-projectionist
The current recurse logic (move up with fnamemodify and :h, check for
root "/") fails if a buffer is opened with a non-UNIX filename. Notably,
this fails for fugitive:// buffers.

This gave insight as to why vim-projectionist is implemented with a root
vs previous check instead of an explicit root check: non-UNIX filenames
will *never* match root "/" and will thus go into an infinite loop.

The same logic has been copied for Makery.